--- chapter01/item01/src/main/java/com/eigpay/idea/Main.java
@@ -1,23 +1,0 @@
-package com.eigpay.idea;
-
-import com.eigpay.idea.assist.GetPerson;
-import com.eigpay.idea.model.Person;
-
-/**
- * description:
- *
- * @author LCN
- * @date 2018-02-20 下午 08:48
- */
-public class Main {
-
-    public static void main(String[] args) {
-        // 通过静态工厂方式获取实例
-        Person student = GetPerson.getStudent();
-        Person teacher = GetPerson.getTeacher();
-
-        student.showJob();
-        teacher.showJob();
-    }
-
-}
