--- chapter01/item01/src/main/java/com/eigpay/idea/demo/MusicProvider.java
@@ -1,11 +1,0 @@
-package com.eigpay.idea.demo;
-
-/**
- * description:
- *
- * @author LCN
- * @date 2018-02-20 下午 08:48
- */
-public interface MusicProvider {
-    MusicApp getMusicApp();
-}
