--- chapter01/item01/src/main/java/com/eigpay/idea/demo/NetEaseMusicProvider.java
@@ -1,21 +1,0 @@
-package com.eigpay.idea.demo;
-
-/**
- * description:
- *
- * @author LCN
- * @date 2018-02-20 下午 08:48
- */
-public class NetEaseMusicProvider implements MusicProvider {
-
-    @Override
-    public MusicApp getMusicApp() {
-        boolean isDeviceAndroid = true;
-        if ( isDeviceAndroid ) {
-            return new NetEaseMusicApp();
-        }
-        System.out.println("无播放器");
-        return null;
-    }
-
-}
