+++ chapter01/item01/src/main/java/com/lcn/idea/demo/MusicApp.java
@@ -1,0 +1,11 @@
+package com.lcn.idea.demo;
+
+/**
+ * description:
+ *
+ * @author LCN
+ * @date 2018-02-20 下午 08:48
+ */
+public interface MusicApp {
+    void play();
+}
