+++ chapter01/item01/src/main/java/com/lcn/idea/demo/NetEaseMusicApp.java
@@ -1,0 +1,15 @@
+package com.lcn.idea.demo;
+
+/**
+ * description:
+ *
+ * @author LCN
+ * @date 2018-02-20 下午 08:48
+ */
+public class NetEaseMusicApp implements MusicApp {
+
+    @Override
+    public void play() {
+        System.out.println("网易云音乐，听见好时光~来自Android客户端");
+    }
+}
