--- chapter01/item02/src/main/java/com/eigpay/idea/demo/Calzone.java
@@ -1,37 +1,0 @@
-package com.eigpay.idea.demo;
-
-/**
- * description: 半圆形烤乳酪馅饼 可以指定是否加酱汁
- *
- * @author LCN
- * @date 2018-02-19 下午 04:19
- */
-public class Calzone extends BasePizza {
-    /** 是否加酱汁 */
-    private final boolean sauceInside;
-
-    public static class Builder extends BasePizza.Builder<Builder> {
-        /** 默认不加酱汁 */
-        private boolean sauceInside = false;
-
-        public Builder sauceInside() {
-            sauceInside = true;
-            return this;
-        }
-
-        @Override
-        public Calzone build() {
-            return new Calzone(this);
-        }
-
-        @Override
-        protected Builder self() {
-            return this;
-        }
-    }
-
-    private Calzone(Builder builder) {
-        super(builder);
-        sauceInside = builder.sauceInside;
-    }
-}
