--- chapter01/item03/src/main/java/com/eigpay/idea/singleton/Three.java
@@ -1,21 +1,0 @@
-package com.eigpay.idea.singleton;
-
-/**
- * description: 实现singleton的方式三 声明单一元素的枚举类
- * 声明单一元素的枚举类这种方式类似于公共属性方法，但更简洁，提供了免费的序列化机制，
- * 并提供了针对多个实例化的坚固保证，即使是在复杂的序列化或反射攻击的情况下。
- * 这种方法可能感觉有点不自然，但是单一元素枚举类通常是实现单例的最佳方式。
- * 注意，如果单例必须继承Enum以外的父类(尽管可以声明一个Enum来实现接口)，那么就不能使用这种方法
- *
- * @author LCN
- * @date 2018-02-20 下午 09:17
- */
-public enum  Three {
-
-    /** 唯一的枚举量，作为单例 */
-    INSTATNCE;
-
-    public void test() {
-        System.out.println("单例实现方式三");
-    }
-}
